Đã fix từ 85 -> 75 ( nhớ check lại sau )

+++ UIT_23520335_Function/Bai070/70.cpp
@@ -2,7 +2,7 @@
 #include<cmath>
 using namespace std;
 
-int tong(int x, int n);
+int tong(float x, int n);
 
 int main()
 {
@@ -12,10 +12,10 @@ int main()
 	return 0;
 }
 
-int tong(int x, int n)
+int tong(float x, int n)
 {
-	int s = 0;
-	int t = 1;
+	float s = 0;
+	float t = 1;
 	int i = 2;
 	while (i <= 2*n)
 	{
